app/requirements.txt
flask==2.2.3
psycopg2-binary==2.9.5
python-dotenv==1.0.0
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
markupsafe==2.1.2
gunicorn==20.1.0
